Using jQuery to smoothly scroll between sections of a page and preserve anchor id in URL
Inclusion of two HTML files to demonstrate one page that scrolls fluidly between sections and preserves the link anchor id in the page URL. The user can go back and forth in their browser and the page will scroll smoothly between the sections. If the user is on a different page in the website and clicks on the same links the animation of scrolling from the top will be preserved.